requirements.txt
bilibili-api-python
pillow
pywin32
requests
curl_cffi
tqdm